# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MatriConnect Matrimony API",
        "description": "Full-stack matrimony platform API. All protected endpoints require Bearer token from Sanctum.",
        "contact": {
            "email": "admin@matriconnect.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/photos/pending": {
            "get": {
                "tags": [
                    "Admin - Photo Moderation"
                ],
                "summary": "Get photos pending moderation",
                "operationId": "f7a420b82eb9105ee4e26e5ff603987f",
                "responses": {
                    "200": {
                        "description": "Pending photos list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/photos/{id}/approve": {
            "put": {
                "tags": [
                    "Admin - Photo Moderation"
                ],
                "summary": "Approve a photo",
                "operationId": "dd98babdcc490e760a381f666c39fae7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Photo approved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/photos/{id}/reject": {
            "put": {
                "tags": [
                    "Admin - Photo Moderation"
                ],
                "summary": "Reject a photo",
                "operationId": "6a7fb0b74941408144d4b6a16d2925a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Inappropriate content"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Photo rejected"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token. Obtain via POST /api/v1/auth/login",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}